fix: satisfy invocation cleanup audit
